[OPMONDEV-171]: code clean up
